A lot of changes, activities added

+++ controllers/team_controller.js
@@ -1,6 +1,8 @@
 'use strict'
 
-var TeamMember = require('../models/team_model');  // Cargo el modelo de Team de la BD
+var TeamMember = require('../models/team_model');   // Cargo el modelo de Team de la BD
+var File = require('../models/file_model');         // Cargo el modelo de archivos de la BD
+var opt = require('../globals');
 var fs = require('fs');             // Librería para el manejo de archivos de node
 var path = require('path');         // método de node que permite acceder a rutas físicas de archivos
 
@@ -31,30 +33,30 @@ var TeamController = {
         }
     },
     // Método para enviar un archivo de imágen al front-end
-    getImageFile: function(req, res){
-        var file = req.params.image;
-        var path_file = './uploads/equipo/'+file;
-        console.log(file);
+    // getImageFile: function(req, res){
+    //     var file = req.params.image;
+    //     var path_file = './uploads/equipo/'+file;
+    //     // console.log(file);
 
-        fs.exists(path_file, (exists) => {
-            if(exists){
-                return res.sendFile(path.resolve(path_file));
-            }
-            else{
-                return res.status(200).send({message: "No existe la imagen..."});
-            }
-        });
-    },
+    //     fs.exists(path_file, (exists) => {
+    //         if(exists){
+    //             return res.sendFile(path.resolve(path_file));
+    //         }
+    //         else{
+    //             return res.status(200).send({message: "No existe la imagen..."});
+    //         }
+    //     });
+    // },
     saveMember: function(req,res){
         var member = new TeamMember();
-        var params = req.body;
-
-        member.name = params.name;
-        member.position = params.position;
-        member.description = params.description;
-        member.picture = null; // La asigno en el método uploadTeamMemberImage
-        member.slider_order = params.slider_order;
-        member.show = params.show;
+        var values = req.body;
+        //console.log("Picture ID: " + values.picture);
+        member.name = values.name;
+        member.position = values.position;
+        member.description = values.description;
+        member.picture = values.picture; 
+        member.slider_order = values.slider_order;
+        member.show = values.show;
 
         member.save((err, TeamMemberStored) => {
             if(err) return status(500).send({message:"Error al guardar el nuevo Miembro"});
@@ -64,75 +66,75 @@ var TeamController = {
             return res.status(200).send({member: TeamMemberStored});
         });
     },
-    uploadTeamMemberImage: function(req,res){
+    // uploadTeamMemberImage: function(req,res){
                 
-        var TeamMemberID = req.params.id;
-        var filePath = req.files.image.path;
-        var originalName = "";
-        originalName = req.files.image.name;
-        var fileName = "No hay imágen!";
-        var fileExtension = "";
-        var extensions = ['png','jpg','jpeg','bmp','png'];
-        var validExtension = false;
+    //     var TeamMemberID = req.params.id;
+    //     var filePath = req.files.image.path;
+    //     var originalName = "";
+    //     originalName = req.files.image.name;
+    //     var fileName = "No hay imágen!";
+    //     var fileExtension = "";
+    //     var extensions = ['png','jpg','jpeg','bmp','png'];
+    //     var validExtension = false;
 
-        if(req.files){
-            fileName = req.files.image.path.split('\equipo\\')[1];
-            fileExtension = fileName.split('.')[1].toLowerCase();
+    //     if(req.files){
+    //         fileName = req.files.image.path.split('\equipo\\')[1];
+    //         fileExtension = fileName.split('.')[1].toLowerCase();
 
-            console.log("originalName: " + originalName);
-            console.log("filePath: " + filePath);
-            console.log("fileName: " + fileName);
-            console.log("fileExtension: " + fileExtension + " Type: " + typeof fileExtension);
+    //         // console.log("originalName: " + originalName);
+    //         // console.log("filePath: " + filePath);
+    //         // console.log("fileName: " + fileName);
+    //         // console.log("fileExtension: " + fileExtension + " Type: " + typeof fileExtension);
             
-            for(var i in extensions){
-                if(fileExtension == extensions[i])
-                {
-                    validExtension = true;
-                }
-            }
+    //         for(var i in extensions){
+    //             if(fileExtension == extensions[i])
+    //             {
+    //                 validExtension = true;
+    //             }
+    //         }
 
-            if(validExtension){
-                var newPath = 'uploads\\equipo\\' + originalName;
-                fs.exists(newPath, (exists) => {
-                    if(exists){
-                        //console.log('El archivo ' + newPath + ' ya existe!');
-                    }
-                    else{
-                        fs.rename(filePath, newPath, (err) => {
-                            if(err){
-                                //console.log('No se pudo cambiar el nombre, se mantiene ' + fileName);
-                                //console.log(err);
-                            }
-                            else{
-                                //console.log("Original: " + fileName);
-                                fileName = originalName;
-                                //console.log("Nuevo: " + fileName);
+    //         if(validExtension){
+    //             var newPath = 'uploads\\equipo\\' + originalName;
+    //             fs.exists(newPath, (exists) => {
+    //                 if(exists){
+    //                     //console.log('El archivo ' + newPath + ' ya existe!');
+    //                 }
+    //                 else{
+    //                     fs.rename(filePath, newPath, (err) => {
+    //                         if(err){
+    //                             //console.log('No se pudo cambiar el nombre, se mantiene ' + fileName);
+    //                             //console.log(err);
+    //                         }
+    //                         else{
+    //                             //console.log("Original: " + fileName);
+    //                             fileName = originalName;
+    //                             //console.log("Nuevo: " + fileName);
 
-                                TeamMember.findByIdAndUpdate(TeamMemberID, {picture: fileName}, {new: true}, (err, memberUpdated) => {
-                                    //console.log("Miembro Actualizado:");
-                                    //console.log(memberUpdated);
+    //                             TeamMember.findByIdAndUpdate(TeamMemberID, {picture: fileName}, {new: true}, (err, memberUpdated) => {
+    //                                 //console.log("Miembro Actualizado:");
+    //                                 //console.log(memberUpdated);
                                     
-                                    if(err) return res.status(500).send({message: 'No se pudo guardar la imagen!'});
+    //                                 if(err) return res.status(500).send({message: 'No se pudo guardar la imagen!'});
                     
-                                    if(!memberUpdated) return res.status(404).send({message: 'No se encuentra el ID!'});
+    //                                 if(!memberUpdated) return res.status(404).send({message: 'No se encuentra el ID!'});
                     
-                                    return res.status(200).send({member: memberUpdated});
-                                });
-                            }
-                        });
-                    }
-                });
-            }
-            else{
-                fs.unlink(filePath, (err) => {  // fs.unlink Borra el archivo subido y envia respuesta
-                    return res.status(200).send({message: 'La extensión no es válida'});
-                });
-            }
-        }
-        else{
-            return res.status(200).send({message: 'No se subió el archivo!'});
-        }
-    },
+    //                                 return res.status(200).send({member: memberUpdated});
+    //                             });
+    //                         }
+    //                     });
+    //                 }
+    //             });
+    //         }
+    //         else{
+    //             fs.unlink(filePath, (err) => {  // fs.unlink Borra el archivo subido y envia respuesta
+    //                 return res.status(200).send({message: 'La extensión no es válida'});
+    //             });
+    //         }
+    //     }
+    //     else{
+    //         return res.status(200).send({message: 'No se subió el archivo!'});
+    //     }
+    // },
     // Método para actualizar un proyecto de la BD
     // Envío un PUT http://localhost:3700/api/update-project/5d812c99f191e22018038323
     // Con los datos a actualizar en body
@@ -154,17 +156,29 @@ var TeamController = {
     deleteMember: function(req,res){
         var memberID = req.params.id;
 
-        TeamMember.findById
-
         TeamMember.findByIdAndDelete(memberID, (err, memberDeleted) => {
 
-            var deleteFile = 'uploads\\equipo\\' + memberDeleted.picture;
-
-            fs.unlink(deleteFile, (error) => {
-                console.log(err);
+            File.findByIdAndDelete(memberDeleted.picture, (err, fileDeleted) => {
+                //console.log(fileDeleted);
+                if(err) return res.status(500).send({message: 'No se ha podido Borrar la imágen de la BD'});
+                if(!fileDeleted){
+                    return res.status(404).send({message: 'No existe la imágen en la BD'});
+                } 
+                else{
+                    var deleteFile = opt.UPLOAD_PATH + '\\' + fileDeleted.filename;
+                    fs.exists(deleteFile, (exists) => {
+                        if(exists){
+                            fs.unlink(deleteFile, (error) => {
+                                if(error) return res.status(500).send({message: 'Error al borrar el archivo de imágen'});
+                                //else return res.status(200).send({file: fileDeleted});
+                            });
+                        }
+                        else return res.status(404).send({message: 'No existe el archivo de imágen en el servidor'});
+                    });
+                }
             });
 
-            if(err) return res.status(500).send({message: 'No se ha podido Borar'});
+            if(err) return res.status(500).send({message: 'No se ha podido Borar el miembro de la BD'});
 
             if(!memberDeleted) return res.status(404).send({message: 'No existe el Miembro'});
